Revert "Move AccountUrls to /account"
